added meaningful components to the project
