+++ src/components/Cockpit/Cockpit.module.css
@@ -1,0 +1,26 @@
+.red {
+  color: red;
+}
+.bold {
+  font-weight: bold;
+}
+.Cockpit button {
+  background-color: green;
+  color: white;
+  font: inherit;
+  border: 1px solid blue;
+  padding: 8px;
+  cursor: pointer;
+}
+
+.Cockpit button:hover {
+  background-color: lightgreen;
+  color: white;
+}
+
+.Cockpit button.Red {
+  background-color: red;
+}
+.COckpit button.Red:hover {
+  background-color: salmon;
+}
