+++ src/components/Persons/Persons.js
@@ -1,5 +1,16 @@
 import React from "react";
+import Person from "./Person/Person";
+const Persons = props =>
+  props.persons.map((person, index) => {
+    return (
+      <Person
+        click={() => props.clicked(index)}
+        name={person.name}
+        age={person.age}
+        key={person.id}
+        changed={event => props.changed(event, person.id)}
+      />
+    );
+  });
 
-export default function Persons() {
-  return <div></div>;
-}
+export default Persons;
