+++ src/containers/App.js
@@ -1,10 +1,9 @@
 import React, { Component } from "react";
-
-import styles from "./App.module.css";
-
-import Person from "../components/Persons/Person/Person";
+import Cockpit from "../components/Cockpit/Cockpit";
+import Persons from "../components/Persons/Persons";
 import Validation from "../components/Validation/Validation";
 import Char from "../components/Char/Char";
+import classes from "./App.module.css";
 
 class App extends Component {
   state = {
@@ -73,43 +72,24 @@ class App extends Component {
     });
 
     let persons = null;
-    let btnClass = [styles.Button];
+
     if (this.state.showPersons) {
       persons = (
-        <div>
-          {this.state.persons.map((person, index) => {
-            return (
-              <Person
-                click={() => this.deletePersonHandler(index)}
-                name={person.name}
-                age={person.age}
-                key={person.id}
-                changed={event => this.nameChangeHandler(event, person.id)}
-              />
-            );
-          })}
-        </div>
+        <Persons
+          persons={this.state.persons}
+          clicked={this.deletePersonHandler}
+          changed={this.nameChangeHandler}
+        />
       );
-      btnClass.push(styles.Red);
     }
-    const assignedClasses = [];
-    if (this.state.persons.length <= 2) {
-      assignedClasses.push(styles.red);
-    }
-    if (this.state.persons.length <= 1) {
-      assignedClasses.push(styles.bold);
-    }
+
     return (
-      <div className="App">
-        <h1>Hi, I'm a react app </h1>
-        <p className={assignedClasses.join(" ")}>This is really working!</p>
-        <button
-          className={btnClass.join(" ")}
-          // alt={this.state.showPersons}
-          onClick={this.togglePersonsHandler}
-        >
-          Toggle Persons
-        </button>
+      <div className={classes.App}>
+        <Cockpit
+          showPersons={this.state.showPersons}
+          persons={this.state.persons}
+          clicked={this.togglePersonsHandler}
+        />
         {persons}
         <br />
         <input
