+++ src/containers/App.module.css
@@ -1,29 +1,3 @@
-body {
+.App {
   text-align: center;
-}
-.red {
-  color: red;
-}
-.bold {
-  font-weight: bold;
-}
-.Button {
-  background-color: green;
-  color: white;
-  font: inherit;
-  border: 1px solid blue;
-  padding: 8px;
-  cursor: pointer;
-}
-
-.Button:hover {
-  background-color: lightgreen;
-  color: white;
-}
-
-.Button.Red {
-  background-color: red;
-}
-.Button.Red:hover {
-  background-color: salmon;
 }
